Don't handle nil actions in navigation middleware
The use case is when you have custom action creator attached to router with "getCustomActionCreators" config property, but it doesn't alway return an action. Example is "handleLink" function which takes "url" argument, validates if this url is a valid deep link and than updates the navigation state, but if url starts with "http" it opens it in WebBrowser and doesn't return any action.